fix(fhirpath): correct JSON value ordering and per-evaluation context isolation
ElementJsonConverter.ToJsonNode checked Meta<JsonNode>() before element.Value, inverting the order JsonNodeMutator.SerializeValue used before the extraction. For a FHIR primitive carrying a _value shadow, Meta<JsonNode>() returns the shadow object, so the converter emitted the extension object and dropped the primitive value - a regression in the mapping engine as well as a bug in instance-selector construction. Restore value-first ordering and document why it matters.

FhirPathInvariantCheck cached a single EvaluationContext and returned it from every unseeded evaluation. EvaluationContext.DefinedVariables is a mutable Dictionary written by defineVariable(), so variables leaked between constraint evaluations and concurrent validation raced on a non-thread-safe dictionary. Cache the creator delegate instead and build a fresh context per evaluation.

Move ElementAssignment to its own file per the one-type-per-file rule.


Copilot-Session: 85de2995-1685-4191-b196-70679aaf7e56
