+++ src/Core/Ignixa.FhirPath/Expressions/InstanceSelectorExpression.cs
@@ -64,24 +64,3 @@ public class InstanceSelectorExpression : Expression
         IFhirPathExpressionVisitor<TContext, TOutput> visitor,
         TContext context) => visitor.VisitInstanceSelector(this, context);
 }
-
-/// <summary>
-/// Represents a single element assignment within an instance selector.
-/// Example: system: 'http://example.org'
-/// </summary>
-public class ElementAssignment
-{
-    public ElementAssignment(string elementName, Expression valueExpression)
-    {
-        ElementName = elementName ?? throw new ArgumentNullException(nameof(elementName));
-        ValueExpression = valueExpression ?? throw new ArgumentNullException(nameof(valueExpression));
-    }
-
-    /// <summary>Name of the element being assigned</summary>
-    public string ElementName { get; }
-
-    /// <summary>Expression that produces the value for this element</summary>
-    public Expression ValueExpression { get; }
-
-    public override string ToString() => $"{ElementName}: {ValueExpression}";
-}
